Update new strategy for solr document not found
When there's no solr document it could be because the resource is gone
but it could also be because the resource was taken down. If it's been
taken down we don't want to delete the sidecar, in case it has custom
data and the resource is put back up.

The 2 strategies now both solve the duplicate sidecars problem and the
results are nearly identical. However, alternative strategy cleans up a
sidecar that holds a reference to a deleted resource, whereas original
strategy will keep that sidecar if it was in a set with sidecars
referencing nil resources, because it will just delete the on with the
nil and then move on.