Add task to delete duplicate sidecars, resources (#942)
* Add task to delete duplicate sidecars, resources

These are blocking upgrade due to a new uniqueness constraint on
sidecars for exhibit_id and document_id.

Co-authored-by: Eliot Jordan <eliot.jordan@gmail.com>

* Add more print statements about updated dates

* A different idea for deleting sidecars

* Update new strategy for solr document not found

When there's no solr document it could be because the resource is gone
but it could also be because the resource was taken down. If it's been
taken down we don't want to delete the sidecar, in case it has custom
data and the resource is put back up.

The 2 strategies now both solve the duplicate sidecars problem and the
results are nearly identical. However, alternative strategy cleans up a
sidecar that holds a reference to a deleted resource, whereas original
strategy will keep that sidecar if it was in a set with sidecars
referencing nil resources, because it will just delete the on with the
nil and then move on.

* Fix some edge cases found in production

* Use "alternate" method. Clean up debug messages.

Co-authored-by: Eliot Jordan <eliot.jordan@gmail.com>

* Use a logger!

Co-authored-by: Eliot Jordan <eliot.jordan@gmail.com>

* Actually destroy data!

Co-authored-by: Eliot Jordan <eliot.jordan@gmail.com>

* Remove ineffective rubocop exclusions

Co-authored-by: Eliot Jordan <eliot.jordan@gmail.com>