Fix override title.
This changes override title to happen at the Blacklight configuration
layer rather than in the presenters. I think this is more what Spotlight
would want you to do, so it's a lot less overrides.

This rewrites a number of the tests to use the real configuration rather
than a bunch of stubs.

Closes #1115