Pull in various style updates from upstream made between 4.0 and 4.6
There's one change I didn't pull through, which added a "restore
default" button next to each metadata field label in the metadata form
in the dashboard. I tried adding this in but it was a little buggy so I
left it out. The PR was https://github.com/projectblacklight/spotlight/pull/3130